requirements.txt
asgiref==3.3.1
Django==3.1.4
django-constance==2.8.0
django-extensions==3.1.0
djangorestframework==3.12.2
gevent==20.9.0
greenlet==0.4.17
psycopg2==2.8.6
python-memcached==1.59
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
zope.event==4.5.0
zope.interface==5.2.0
